pyproject.toml
[tool.poetry]
name = "confident"
version = "0.2.2"
description = "Loading configurations from multiple sources into a data model."
authors = ["limonyellow"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
pydantic = "^1.9.0"
pyyaml = "^6.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
